+++ modelo/src/main/java/cl/madal/skorpion/repository/ProvinciaRepository.java
@@ -1,8 +1,3 @@
-/*
- * To change this license header, choose License Headers in Project Properties.
- * To change this template file, choose Tools | Templates
- * and open the template in the editor.
- */
 package cl.madal.skorpion.repository;
 
 import cl.madal.skorpion.modelo.Provincia;
@@ -15,7 +10,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
  * @author Juan
  */
 @Resource(name = "provinciaRepository")
-public interface ProvinciaRepository extends JpaRepository<Provincia, Integer> {
+public interface ProvinciaRepository extends JpaRepository<Provincia, Long> {
 
     public Provincia findByNombreIgnoreCaseAndRegion(String nombre, Region region);
 }
