+++ modelo/src/main/java/cl/madal/skorpion/repository/RegionRepository.java
@@ -9,11 +9,10 @@ import org.springframework.data.jpa.repository.JpaRepository;
  * @author Sebastián Salazar Molina <sebasalazar@gmail.com>
  */
 @Resource(name = "regionRepository")
-public interface RegionRepository extends JpaRepository<Region, Integer> {
+public interface RegionRepository extends JpaRepository<Region, Long> {
 
     public Region findByNombreIgnoreCase(String nombre);
-    
-      public Region findByNumero(Integer numero);
-    
-    
+
+    public Region findByNumero(Integer numero);
+
 }
